Fixed little inconvenience in empty inputs.

+++ app/Http/Controllers/PostController.php
@@ -11,6 +11,8 @@ use App\Models\User;
 use App\Models\Content;
 use App\Models\Category;
 
+use Illuminate\Support\Facades\Input;
+
 class PostController extends Controller
 {
   /**
@@ -79,6 +81,11 @@ class PostController extends Controller
    */
   public function create(Request $request)
   {
+    $categories = array_filter(explode(',', $request->input('categories')));
+
+    if (empty($categories) || $request->input('title') !== '' || $request->input('body') !== '')
+      return redirect()->back()->withInput(Input::all());
+
     //$this->authorize('create', $post);
     
     $content = new Content;
@@ -93,8 +100,6 @@ class PostController extends Controller
     $post->title = $request->input('title');
     $post->save();
 
-    $categories = explode(',', $request->input('categories'));
-
     DB::table('post_category')->where('post', $post->id)->delete();
 
     foreach ($categories as $category) {
@@ -108,11 +113,14 @@ class PostController extends Controller
 
   public function edit(Request $request, $id)
   {
+    $categories = array_filter(explode(',', $request->input('categories')));
+
+    if(empty($categories) || $request->input('title') !== '' || $request->input('body') !== '')
+      return redirect()->back()->withInput(Input::all());
+
     $post = Post::find($id);
 
     //$this->authorize('edit', $post);
-
-    $categories = explode(',', $request->input('categories'));
 
     DB::table('post_category')->where('post', $id)->delete();
